requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.9.4
Jinja2==2.7.1
MarkupSafe==0.18
MySQL-python==1.2.4
WTForms==1.0.5
Werkzeug==0.9.6
BeautifulSoup==3.2.1
requests==2.2.1
gunicorn==19.7.1
PyMySQL==0.7.11
